Made changes to the robot class and renamed it appropriately
